Remove unnecessary boxing and unboxing

+++ src/main/java/com/avaje/ebeaninternal/server/type/LongToTimestampConverter.java
@@ -12,7 +12,7 @@ public class LongToTimestampConverter implements ScalarTypeConverter<Long, Times
 
   public Timestamp unwrapValue(Long beanType) {
 
-    return new Timestamp(beanType.longValue());
+    return new Timestamp(beanType);
   }
 
   public Long wrapValue(Timestamp scalarType) {
